pre-final fix v2

+++ src/main/java/webapp/Calc.java
@@ -38,11 +38,11 @@ public class Calc extends HttpServlet {
 			
 			request.setAttribute("radius", radius_calc);
 			request.setAttribute("height", height_calc);
-			int radius_try;
-			int height_try;
+			double radius_try;
+			double height_try;
 			try {
-				radius_try=Integer.parseInt(radius_calc);
-				height_try=Integer.parseInt(height_calc);
+				radius_try=Double.parseDouble(radius_calc);
+				height_try=Double.parseDouble(height_calc);
 			} catch (NumberFormatException e) {
 				radius_try=0;
 				height_try=0;
